fix: large files broke prefetch
backport master commit 7e39a5d8f1a7dc3555ce474d5e6b0c76c5df7e79
Author: 泰友 <cuichengxu.ccx@antgroup.com>
Date:   Wed Jul 5 14:14:09 2023 +0800

    fix: large files broke prefetch

    Files larger than 4G leads to prefetch panic, because the max blob io
    range is smaller than 4G. This pr changes blob io max size from u32 to
    u64.

+++ rafs/src/metadata/layout/v5.rs
@@ -1324,7 +1324,7 @@ fn add_chunk_to_bio_desc(
         (chunk_end - chunk_start) as u32,
         user_io,
     );
-    desc.bi_size += bio.size;
+    desc.bi_size += bio.size as u64;
     desc.bi_vec.push(bio);
 
     true
